Added unit test for charlieplexed LEDs
The 2x2 LEDs are charlieplexed meaning that
they are easily addressed one by one but
cannot be turned on together
In order to turn them on together, they
need to be flickered. Meaning they have
to be turned off and on quickly to make
it look like they are being turned on
simultaneously.
The results are quite good for 4 LEDs

Signed-off-by: nisha <nishakumarx@gmail.com>