Added different loops for different number of LEDs
From observation it seems that pinMode setup takes up
a lot of time and increases based on how many LEDs
are turned on. Turning on 1 LED of course is the
fastest by far.

There is a Serial console statement to tune the delay
based on the number of LEDs. These should be removed
once tuning is done as it seems to affect the
responsiveness of the LEDs when they flicker
It doesn't look that good

Signed-off-by: nisha <nishakumarx@gmail.com>