Removed cap and replaced JST connector with vias
VPP on the UART is for a one time ROM programming of the UART
which we don't need, so removing it.

We were unable to find batteries that came with the JST female
connectors that went with the used board mounted connector.
Replacing it with vias to use with either board mounted breakaway
header or something else which will fit.

Signed-off-by: nisha <nishakumarx@gmail.com>